embedding-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# Machine Learning
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
open-clip-torch==2.24.0
Pillow==10.1.0
numpy==1.24.3

# File Processing - NEW!
pdfplumber==0.10.0
pandas==2.1.4
python-magic==0.4.27
openpyxl==3.1.2

# RAG & LLM - NEW!
langchain==0.1.0
openai==1.3.8

# Analytics & Visualization - NEW!
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.0
aiohttp==3.9.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
